image_from_a_URL_to_sound/requirements.txt
Pillow==8.3.2
gTTS==2.2.2
pytesseract==0.3.8
pytest==6.2.5
